+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## 3.0.0 (unreleased)
+
+Rework Emoji support:
+
+- Only reduce Emoji width of RGI Emoji (configurable)
+- VS16 turns Emoji characters of width 1 into full-width
+- Add option to make Text Presentation Emoji full-width
+- Emoji widths are now enabled by default
+
 ## 2.6.0
 
 - Unicode 16